% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{accessible_buckets}
\alias{accessible_buckets}
\title{Get a vector of the buckets (base folders) which the user has permission to read from}
\usage{
accessible_buckets()
}
\value{
character
}
\description{
Get a vector of the buckets (base folders) which the user has permission to read from
}
\examples{
accessible_buckets()
}
