% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_helper_functions.R
\name{accessible_files_df}
\alias{accessible_files_df}
\title{Return a dataframe all accessable files, including full path and filesize information
Note:  The listing includes folders as well as files - this enables us to derive a recursive dir structure}
\usage{
accessible_files_df(bucket_filter = NULL)
}
\arguments{
\item{bucket_filter}{return only buckets that match this character vector of bucket names}
}
\value{
data frame (tbl_df) with path of all files available to you in S3.
}
\description{
Return a dataframe all accessable files, including full path and filesize information
Note:  The listing includes folders as well as files - this enables us to derive a recursive dir structure
}
\examples{
accessible_files_df()

}
