% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_utils.R
\name{fetch_paws_creds}
\alias{fetch_paws_creds}
\title{Credentials for Jira API}
\usage{
fetch_paws_creds(type = "", aws_profile = "", region = "")
}
\arguments{
\item{type}{public or private type of access, if private provide profile name.}

\item{aws_profile}{name of profile to use from ~/.aws/config}

\item{region}{region of the s3 bucket, defaults to us-west-2}
}
\value{
credentials required for access to s3 resources based on the supplied aws profile, ex. janus or customized profile
}
\description{
Credentials for Jira API
}
