% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_helper_functions.R
\name{find_s3_paths}
\alias{find_s3_paths}
\title{Search all available files, return a vector of paths}
\usage{
find_s3_paths(search_pattern)
}
\value{
a vector of matching paths
}
\description{
Search all available files, return a vector of paths
}
\examples{
find_s3_paths('temp\\\\d')

}
