% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Get AWS credentials}
\usage{
get_credentials(key = NULL, secret = NULL, token = NULL, ...)
}
\arguments{
\item{key}{A character string specifying an AWS Access Key ID (optional)}

\item{secret}{A character string specifying an AWS Secret Access Key (optional)}

\item{token}{A character string specifying an AWS Session Token (optional)}

\item{...}{Additional arguments passed to \code{credentials_resolver}}
}
\description{
Get AWS credentials from the first available source (passed arguments,
environment variables, specified file, default file, ec2 metadata iam role)
}
\examples{
\dontrun{
# build a credentials object
get_credentials(key = "AWS access key id", secret = "AWS secret access key")

# fetch credentials from specified credentials file
get_credentials(filename = "credentials-file", profile = "default")

# fetch credentials from environment variables, or default credentials file,
# or IAM role from EC2 metadata, respectively
get_credentials()

# As above, but skip environment variables and fetch credentials for
# specified profile name
get_credentials(profile = "default")
}
}
