% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshable_credentials.R
\name{is.credentials}
\alias{is.credentials}
\title{Is the object an AWS credentials object?}
\usage{
is.credentials(x)
}
\description{
Is the object an AWS credentials object?
}
\keyword{internal}
