% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshable_credentials.R
\name{is.refreshable_credentials}
\alias{is.refreshable_credentials}
\title{Is the object a refreshable AWS credentials object?}
\usage{
is.refreshable_credentials(x)
}
\description{
Is the object a refreshable AWS credentials object?
}
\keyword{internal}
