% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing_utils.R
\name{list_s3_files}
\alias{list_s3_files}
\title{List files in a given bucket + prefix that matches a specific pattern, with pagination + option to specify max keys searched}
\usage{
list_s3_files(bucket_name, file_pattern, prefix = "", s3_init, max_keys = 1000)
}
\arguments{
\item{bucket_name}{name of the bucket without s3://}

\item{file_pattern}{name of the bucket without s3://}

\item{prefix}{prefix path to search, ex studyid/ngs/outdata}

\item{s3_init}{output of fetch_paws_creds or paws::s3() to use for credentials}

\item{max_keys}{numeric, number of keys to pass to MaxKeys in list_objects_v2. Default = 1000}
}
\value{
set logic for rest of function, if s3 or local loading should be used
}
\description{
List files in a given bucket + prefix that matches a specific pattern, with pagination + option to specify max keys searched
}
