% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{location_check}
\alias{location_check}
\title{Check if files are in s3 or on prem, used within other functions}
\usage{
location_check(char_vec_files)
}
\arguments{
\item{char_vec_files}{list of file locations}
}
\value{
set logic for rest of function, if s3 or local loading should be used
}
\description{
Check if files are in s3 or on prem, used within other functions
}
