% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-helper-functions.R
\name{moj.s3.dir}
\alias{moj.s3.dir}
\title{Return path of all accessable files}
\usage{
moj.s3.dir()
}
\value{
data frame with path of all files available to you
}
\description{
Return path of all accessable files
}
\examples{
moj.s3.dir()

}
