% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_utils.R
\name{paws_readRDS}
\alias{paws_readRDS}
\title{Read a serialized file (.rds) from s3}
\usage{
paws_readRDS(bucket_name, object_key, s3_init)
}
\arguments{
\item{bucket_name}{name of the bucket without s3://}

\item{object_key}{key for object to be read}

\item{s3_init}{output of paws::s3() to use for credentials}
}
\value{
object in format of the function used
}
\description{
Read a serialized file (.rds) from s3
}
