% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_utils.R
\name{paws_readusing}
\alias{paws_readusing}
\title{Read a non-serialized file from s3 with a specified function}
\usage{
paws_readusing(
  bucket_name,
  object_key,
  read_function,
  s3_init,
  read_args = list()
)
}
\arguments{
\item{bucket_name}{name of the bucket without s3://,}

\item{object_key}{key for object to be read}

\item{read_function}{function to use for reading, ex read.delim}

\item{s3_init}{output of paws::s3() to use for credentials}

\item{read_args}{optional, additional arguments used by the read_function if needed}
}
\value{
object in format of the function used
}
\description{
Read a non-serialized file from s3 with a specified function
}
