% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_utils.R
\name{paws_writeusing}
\alias{paws_writeusing}
\title{Write object to s3 using a specific function}
\usage{
paws_writeusing(
  object,
  bucket_name,
  object_key,
  write_function,
  write_args = list(),
  s3_init
)
}
\arguments{
\item{object}{R object to be saved}

\item{bucket_name}{name of the bucket without s3://}

\item{object_key}{function to use for writing, ex saveRDS or write.csv}

\item{write_function}{location of inputs for the batch}

\item{write_args}{optionsl, list of args to pass to the write_function ex row.names=F}

\item{s3_init}{output of paws::s3() to use for credentials}
}
\value{
Message of these locations match, prevents improper runs
}
\description{
Write object to s3 using a specific function
}
