% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshable_credentials.R
\name{refresh}
\alias{refresh}
\title{Refresh AWS credentials}
\usage{
refresh(creds)
}
\arguments{
\item{creds}{A credentials object}
}
\description{
Refresh a set of AWS credentials, if possible
}
