% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshable_credentials.R
\name{refresh.refreshable_credentials}
\alias{refresh.refreshable_credentials}
\title{Refresh AWS credentials}
\usage{
\method{refresh}{refreshable_credentials}(creds)
}
\arguments{
\item{creds}{A credentials object}
}
\description{
Refresh a set of AWS credentials, if possible
}
