% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-helper-functions.R
\name{s3.buckets}
\alias{s3.buckets}
\title{Finds out what buckets are available to you}
\usage{
s3.buckets(accessible = TRUE)
}
\arguments{
\item{accessible}{Logial, accessible buckets only? FALSE to see all}
}
\value{
a data frame of buckets you have access to
}
\description{
Finds out what buckets are available to you
}
\examples{
moj.s3.buckets()

}
