% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-helper-functions.R
\name{s3.dir}
\alias{s3.dir}
\title{A directory function for s3}
\usage{
s3.dir(current_path = "")
}
\arguments{
\item{current_path}{a string with the path of the folder to query}
}
\value{
list of directories
}
\description{
A directory function for s3
}
\examples{
s3.dir('directory')
}
