% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-helper-functions.R
\name{s3.fetch.files}
\alias{s3.fetch.files}
\title{Return path of all accessable files}
\usage{
s3.fetch.files()
}
\value{
data frame (tbl_df) with path of all files available to you in S3.
}
\description{
Return path of all accessable files
}
\examples{
moj.s3.dir()

}
