% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_helper_functions.R
\name{s3_dir}
\alias{s3_dir}
\title{A directory function for s3
Returns the contents of the given path, including files and directories}
\usage{
s3_dir(current_path = NULL)
}
\arguments{
\item{current_path}{a string with the path of the folder to query}
}
\value{
list of directories
}
\description{
A directory function for s3
Returns the contents of the given path, including files and directories
}
\examples{
s3_dir('directory')
}
