% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_s3_data.R
\name{s3_path_to_full_df}
\alias{s3_path_to_full_df}
\title{Read a full file from S3, using the full path to the file including the bucketname}
\usage{
s3_path_to_full_df(path)
}
\arguments{
\item{path}{a string -  the full path to the file including the bucketname}
}
\value{
a tibble (dataframe)
}
\description{
Read a full file from S3, using the full path to the file including the bucketname
}
\examples{
df <- s3_read_path_to_df("alpha-moj-analytics-scratch/a/b/c/robins_temp.csv")
}
