% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_s3_data.R
\name{search_and_preview_dfs}
\alias{search_and_preview_dfs}
\title{Preview the first few records of any dataset that matches a search}
\usage{
search_and_preview_dfs(pattern, maxreturns = 10)
}
\arguments{
\item{pattern}{a string -  a regular expression representing the search}
}
\value{
a list of tibbles
#' @export
}
\description{
Preview the first few records of any dataset that matches a search
}
\examples{
 search_and_preview_dfs('temp\\\\d')
}
