% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aws_credentials_env_vars.R
\name{set_aws_credentials_env_vars}
\alias{set_aws_credentials_env_vars}
\title{Set AWS credentials env vars from credentials object}
\usage{
set_aws_credentials_env_vars(creds)
}
\arguments{
\item{creds}{A credentials or refreshable_credentials object}
}
\description{
Set AWS credentials env vars from credentials object
}
