% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_s3.R
\name{write_df_to_csv_in_s3}
\alias{write_df_to_csv_in_s3}
\title{Write an in-memory data frame to a csv file stored in S3}
\usage{
write_df_to_csv_in_s3(df, s3_path, overwrite = FALSE, multipart = TRUE,
  ...)
}
\arguments{
\item{df}{A data frame you want to upload to S3.}

\item{s3_path}{A character string containing the full path to where the file should be stored in S3, including any directory names and the bucket name.}

\item{overwrite}{A logical indicating whether to overwrite the file if it already exists.}

\item{multipart}{A logical indicating whether to use multipart uploads. See \url{http://docs.aws.amazon.com/AmazonS3/latest/dev/mpuoverview.html}. If \code{df} is less than 100 MB when written to csv, this is ignored.}
}
\value{
Returns nothing
}
\description{
Write an in-memory data frame to a csv file stored in S3
}
\examples{
write_df_to_csv_in_s3(df, "alpha-everyone/delete/my_csv.csv")
}
