% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_s3.R
\name{write_df_to_table_in_s3}
\alias{write_df_to_table_in_s3}
\title{Write an in-memory data frame to a csv file stored in S3}
\usage{
write_df_to_table_in_s3(df, s3_path, overwrite = FALSE,
  multipart = TRUE, sep = ",", ...)
}
\arguments{
\item{df}{A data frame you want to upload to S3.}

\item{s3_path}{A character string containing the full path to where the file should be stored in S3, including any directory names and the bucket name.}

\item{overwrite}{A logical indicating whether to overwrite the file if it already exists.}

\item{multipart}{A logical indicating whether to use multipart uploads. See \url{http://docs.aws.amazon.com/AmazonS3/latest/dev/mpuoverview.html}. If \code{df} is less than 100 MB when written to csv, this is ignored.}

\item{sep}{A string used to separate values within each row of \code{df}.}

\item{...}{Additional arguments passed to \code{write.table}.}
}
\value{
Returns nothing
}
\description{
This function is similar to \code{\link{write_df_to_csv_in_s3}} but uses \code{\link[utils]{write.table}} to write the data frame to a csv as opposed to \code{\link[utils]{write.csv}}. This allows the following additional arguments to be passed to \code{\link[utils]{write.table}}: \code{append}, \code{col.names}, \code{sep}, \code{dec}, \code{qmethod}.
}
\examples{
write_df_to_table_in_s3(df, "alpha-everyone/delete/my_csv.csv")
}
