% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_s3.R
\name{write_file_to_s3}
\alias{write_file_to_s3}
\title{Write a file to s3}
\usage{
write_file_to_s3(local_file_path, s3_path, overwrite = FALSE)
}
\arguments{
\item{local_file_path}{- the file you want to upload to s3}

\item{overwrite}{boolean - overwrite the file if it already exists}

\item{s3_file_path}{a string -  the full path of where you want to store the file in s3, including any directory names, but excluding the bucket name}

\item{bucket}{a string - the name of the bucket you want to store the file in}
}
\value{
Returns nothing
}
\description{
Write a file to s3
}
\examples{
local_file_path("myfiles/mydata.csv", "alpha-everyone/delete/my_csv.csv")
}
