% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_helper_functions.R
\name{accessible_buckets}
\alias{accessible_buckets}
\title{Finds out what buckets are available to you}
\usage{
accessible_buckets(accessible = TRUE)
}
\arguments{
\item{accessible}{Logial, accessible buckets only? FALSE to see all}
}
\value{
a character vector of buckets you have access to
}
\description{
Finds out what buckets are available to you
}
\examples{
s3.buckets()

}
