% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-helper-functions.R
\name{accessible_files_df}
\alias{accessible_files_df}
\title{Return a dataframe all accessable files, including full path and filesize information}
\usage{
accessible_files_df()
}
\value{
data frame (tbl_df) with path of all files available to you in S3.
}
\description{
Return a dataframe all accessable files, including full path and filesize information
}
\examples{
accessible_files_df()

}
