% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_helper_functions.R
\name{list_files_in_buckets}
\alias{list_files_in_buckets}
\title{Return a dataframe of accessable files, including full path and filesize information
Note:  The listing includes folders as well as files}
\usage{
list_files_in_buckets(bucket_filter = NULL, prefix = NULL,
  path_only = FALSE)
}
\arguments{
\item{bucket_filter}{return only buckets that match this character vector of bucket names e.g. "alpha-everyone"}

\item{prefix}{filter files which begin with this prefix e.g. 'my_folder/'}

\item{path_only}{boolean - return the accessible paths only, as a character vector}
}
\value{
data frame with path of all files available to you in S3.
}
\description{
Return a dataframe of accessable files, including full path and filesize information
Note:  The listing includes folders as well as files
}
\examples{
accessible_files_df()

}
