% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-helper-functions.R
\name{moj.s3.dir}
\alias{moj.s3.dir}
\title{A directory function for s3}
\usage{
moj.s3.dir(current.path = "")
}
\arguments{
\item{current.path}{a string with the path of the folder to query}
}
\value{
list of directories
}
\description{
A directory function for s3
}
\examples{
moj.s3.dir('directory')
}
