% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_parse_methods.R
\name{s3_path_to_df}
\alias{s3_path_to_df}
\title{Read data from s3 with automatic processing}
\usage{
s3_path_to_df(s3_path, ...)
}
\arguments{
\item{...}{arguemtns passed to read.csv or read_excel.}

\item{path}{path to the s3 file bucket/folder/file.csv}
}
\value{
dataframe
}
\description{
Currently supports csv, tsv, xls, and xlsx files.
}
\examples{
s3tools::s3_path_to_full_df("alpha-test-team/mpg.csv")
s3tools::s3_path_to_preview_df("alpha-test-team/mpg.csv")
}
