% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_s3.R
\name{write_df_to_csv_in_s3}
\alias{write_df_to_csv_in_s3}
\title{Write an in-memory dataframe to a csv file stored in S3}
\usage{
write_df_to_csv_in_s3(df, s3_path, overwrite = FALSE, ...)
}
\arguments{
\item{df}{- the dataframe you want to store in s3}

\item{overwrite}{boolean - overwrite the file if it already exists}

\item{file_path}{a string -  the full path of where you want to store the file in s3, including any directory names, but excluding the bucket}

\item{bucket}{a string - the name of the bucket you want to store the file in}
}
\value{
Returns nothing
}
\description{
Write an in-memory dataframe to a csv file stored in S3
}
\examples{
write_df_to_csv_in_s3(df, "alpha-everyone/delete/my_csv.csv")
}
