% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_s3.R
\name{write_file_to_s3}
\alias{write_file_to_s3}
\title{Write a file to S3}
\usage{
write_file_to_s3(local_file_path, s3_path, overwrite = FALSE,
  multipart = TRUE)
}
\arguments{
\item{local_file_path}{A character string containing the filename (or full file path) of the file you want to upload to S3.}

\item{s3_path}{A character string containing the full path to where the file should be stored in S3, including any directory names and the bucket name.}

\item{overwrite}{A logical indicating whether to overwrite the file if it already exists.}

\item{multipart}{A logical indicating whether to use multipart uploads. See \url{http://docs.aws.amazon.com/AmazonS3/latest/dev/mpuoverview.html}. If the file specified by \code{local_file_path} is less than 100 MB, this is ignored.}
}
\value{
Returns nothing
}
\description{
Write a file to S3
}
\examples{
write_file_to_s3("myfiles/mydata.csv", "alpha-everyone/delete/my_csv.csv")
}
